requirements.txt
# Relaxed constraints to avoid pip resolution conflicts on Render
Flask>=3.0,<4
line-bot-sdk>=3.8,<4
google-api-python-client>=2.120,<3
google-auth>=2.28,<3
python-dotenv>=1.0,<2
requests>=2.31,<3
gunicorn>=21,<23
